lessons/django/solution6/car_admin/requirements.txt
Django==3.0.3
django-ckeditor==5.9.0
pillow==7.0.0
